Sounds and City screen draft
Added a sound for button presses
Probably should create a base button class where all buttons in the game inherit from it.
Added an image to the city screen.
Player HUD now collapses when it is in a city/planet, it shows up again when player starts moving away from city

BUG - Don't know why but buttons in the city screen don't work when player reaches a city that he visited before