FIX: Buoys Percentage, Buoys on current path only
Updated the buoys, it can now be controlled with a percentage on how many you want.

They now also dont reappear after reaching the spacestation on the other path.

Minor changes for the skill check and also corrected the players starting speed and fov.

Added a BP experimental sun

Details Reenabled just for testing if it fits